pyproject.toml
[project]
name = "reddit-scraper"
version = "0.1.0"
description = "Reddit scraper with Apify, FastAPI, and Neo4j graph database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.5.0",
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "neo4j>=5.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
]
